feat: move default configuration to separate module, enhance validation for review prompt and API keys
